+++ python/src/scipp/plot/plot_3d.py
@@ -5,19 +5,19 @@
 # Scipp imports
 from .. import config
 from .render import render_plot
+from ..plot.sciplot import SciPlot
 from .slicer import Slicer
-from .._utils import name_with_unit
+from .._utils import name_with_unit, value_to_string
+from .._scipp import core as sc
 
 # Other imports
 import numpy as np
 import ipywidgets as widgets
 from matplotlib import cm
-
-try:
-    import ipyvolume as ipv
-    from ipyevents import Event
-except ImportError:
-    ipv = None
+import matplotlib as mpl
+from matplotlib.backends import backend_agg
+import PIL as pil
+import pythreejs as p3
 
 
 def plot_3d(scipp_obj_dict=None,
@@ -30,19 +30,15 @@ def plot_3d(scipp_obj_dict=None,
             log=False,
             vmin=None,
             vmax=None,
-            color=None):
+            color=None,
+            background="#f0f0f0"):
     """
-    Plot a 3-slice through a N dimensional dataset. For every dimension above
-    3, a slider is created to adjust the position of the slice in that
-    particular dimension. For other dimensions, the sliders are used to adjust
-    the position of the slice in 3D space.
+    Plot a 3D point cloud through a N dimensional dataset.
+    For every dimension above 3, a slider is created to adjust the position of
+    the slice in that particular dimension.
+    It is possible to add cut surfaces as cartesian, cylindrical or spherical
+    planes.
     """
-
-    # Protect against unloaded module
-    if ipv is None:
-        raise RuntimeError("Three-dimensional projections require ipyvolume "
-                           "and ipyevents to be installed. Use conda/pip "
-                           "install ipyvolume ipyevents.")
 
     sv = Slicer3d(scipp_obj_dict=scipp_obj_dict,
                   axes=axes,
@@ -52,11 +48,12 @@ def plot_3d(scipp_obj_dict=None,
                   vmin=vmin,
                   vmax=vmax,
                   color=color,
-                  aspect=aspect)
+                  aspect=aspect,
+                  background=background)
 
-    render_plot(widgets=sv.box, filename=filename, ipv=ipv)
+    render_plot(widgets=sv.box, filename=filename)
 
-    return sv.members
+    return SciPlot(sv.members)
 
 
 class Slicer3d(Slicer):
@@ -69,7 +66,8 @@ class Slicer3d(Slicer):
                  vmin=None,
                  vmax=None,
                  color=None,
-                 aspect=None):
+                 aspect=None,
+                 background=None):
 
         super().__init__(scipp_obj_dict=scipp_obj_dict,
                          axes=axes,
@@ -82,292 +80,513 @@ class Slicer3d(Slicer):
                          aspect=aspect,
                          button_options=['X', 'Y', 'Z'])
 
-        self.cube = None
-        self.members.update({"surfaces": {}, "wireframes": {}})
+        self.vslice = None
+        self.current_cut_surface_value = None
+        self.cut_slider_steps = 50.
+        self.cbar_image = widgets.Image()
 
-        # Initialise Figure and VBox objects
-        self.fig = ipv.figure(width=config.plot.width,
-                              height=config.plot.height,
-                              animation=0)
-
+        # Prepare colormaps
         self.scalar_map = cm.ScalarMappable(
             norm=self.params["values"][self.name]["norm"],
             cmap=self.params["values"][self.name]["cmap"])
+        self.masks_scalar_map = None
+        if self.params["masks"][self.name]["show"]:
+            self.masks_scalar_map = cm.ScalarMappable(
+                norm=self.params["values"][self.name]["norm"],
+                cmap=self.params["masks"][self.name]["cmap"])
 
+        # Generate the colorbar image
+        self.create_colorbar()
+
+        # Useful variables
         self.permutations = {"x": ["y", "z"], "y": ["x", "z"], "z": ["x", "y"]}
+        self.remaining_inds = [0, 1]
 
-        # Store min/max for each dimension for invisible scatter
-        self.xminmax = dict()
-        for dim, var in self.slider_x[self.name].items():
-            self.xminmax[dim] = [var.values[0], var.values[-1]]
-        self.set_axes_range()
+        # Search the coordinates to see if one contains vectors. If so, it will
+        # be used as position vectors.
+        self.positions = None
+        self.pixel_size = 1.0
+        for coord in self.data_array.coords.values():
+            if coord.dtype == sc.dtype.vector_3_float64:
+                self.positions = np.array(coord.values, dtype=np.float32)
+                break
+        # If no positions are found, create a meshgrid from coordinate axes.
+        if self.positions is None:
+            coords = []
+            for dim, val in self.slider.items():
+                if val.disabled:
+                    coords.append(self.slider_x[self.name][dim].values)
+            z, y, x = np.meshgrid(*coords, indexing='ij')
+            self.positions = np.array(
+                [x.ravel(), y.ravel(), z.ravel()], dtype=np.float32).T
+            self.pixel_size = coords[0][1] - coords[0][0]
 
-        self.wireframes = dict()
-        self.surfaces = dict()
-
-        # #====================================================================
-        # wframes = self.get_outlines()
-        # meshes = self.get_meshes()
-        # surf_args = dict.fromkeys(self.permutations)
-        # wfrm_args = dict.fromkeys(self.permutations)
-        # # print(wframes)
-        # for xyz, perm in self.permutations.items():
-        #     print(xyz, perm)
-        #     key = self.button_axis_to_dim[xyz]
-
-        #     wfrm_args[xyz] = np.ones_like(wframes[xyz][perm[0]]) * \
-        #         self.slider_x[key].values[self.slider[key].value]
-        #     surf_args[xyz] = np.ones_like(meshes[xyz][perm[0]]) * \
-        #         self.slider_x[key].values[self.slider[key].value]
-        #     for p in perm:
-        #         wfrm_args[p] = wframes[xyz][p]
-        #         surf_args[p] = meshes[xyz][p]
-
-        #     self.wireframes[xyz] = ipv.plot_wireframe(**wfrm_args,
-        #                                               color="red")
-        #     self.surfaces[xyz] = ipv.plot_surface(**surf_args, color="red")
-        # #====================================================================
-
-        self.mouse_events = dict()
-        self.last_changed_slider_dim = None
-        for dim, sl in self.slider.items():
-            self.mouse_events[dim] = Event(source=sl,
-                                           watched_events=['mouseup'])
-            self.mouse_events[dim].on_dom_event(self.update_surface)
-
-        # Call update_slice once to make the initial image
-        self.update_axes()
-        self.box = [ipv.gcc()] + self.vbox
-        self.box = widgets.VBox(self.box)
-        self.box.layout.align_items = 'center'
-
-        self.members["fig"] = self.fig
-
-        return
-
-    def update_buttons(self, owner, event, dummy):
-        for dim, button in self.buttons.items():
-            if (button.value == owner.value) and (dim != owner.dim):
-                button.value = owner.old_value
-                button.old_value = button.value
-        owner.old_value = owner.value
-        # Show all surfaces, hide all wireframes
-        for key in self.surfaces.keys():
-            self.surfaces[key].visible = True
-            self.wireframes[key].visible = False
-        # Update the show/hide checkboxes
-        for dim, button in self.buttons.items():
-            ax_dim = button.value
-            if ax_dim is not None:
-                ax_dim = ax_dim.lower()
-            self.showhide[dim].value = (button.value is not None)
-            self.showhide[dim].disabled = (button.value is None)
-            self.showhide[dim].description = "hide"
-            if button.value is None:
-                self.showhide[dim].button_style = ""
-            else:
-                self.showhide[dim].button_style = "success"
-                self.button_axis_to_dim[ax_dim] = dim
-
-        self.fig.meshes = []
-        self.set_axes_range()
-        self.update_axes()
-        # self.box.children = tuple([ipv.gcc()] + self.vbox)
-
-        return
-
-    def update_axes(self):
-        # Go through the buttons and select the right coordinates for the axes
-        titles = dict()
-        buttons_dims = {"x": None, "y": None, "z": None}
-        for dim, button in self.buttons.items():
-            if button.value is not None:
-                titles[button.value.lower()] = name_with_unit(
-                    self.slider_x[self.name][dim], name=str(dim))
-                # buttons_dims[button.value.lower()] = button.dim
-                buttons_dims[button.value.lower()] = dim
-
-        self.fig.xlabel = titles["x"]
-        self.fig.ylabel = titles["y"]
-        self.fig.zlabel = titles["z"]
-
-        self.update_cube()
-
-        return
-
-    def update_cube(self, update_coordinates=True):
-        # The dimensions to be sliced have been saved in slider_dims
-        self.cube = self.data_array
-        self.last_changed_slider_dim = None
-        # Slice along dimensions with buttons who have no value, i.e. the
-        # dimension is not used for any axis. This reduces the data volume to
-        # a 3D cube.
-        for dim, val in self.slider.items():
-            if self.buttons[dim].value is None:
-                self.lab[dim].value = self.make_slider_label(
-                    self.slider_x[self.name][dim], val.value)
-                self.cube = self.cube[dim, val.value]
-
-        # The dimensions to be sliced have been saved in slider_dims
-        button_dim = dict()
-        vslices = dict()
-        # Slice along dimensions with sliders who have a button value
-        for dim, val in self.slider.items():
-            if self.buttons[dim].value is not None:
-                s = self.buttons[dim].value.lower()
-                button_dim[s] = dim
-                self.lab[dim].value = self.make_slider_label(
-                    self.slider_x[self.name][dim], val.value)
-                vslices[s] = {
-                    "slice": self.cube[dim, val.value],
-                    "loc": self.slider_x[self.name][dim].values[val.value]
-                }
-
-        # Now make 3 slices
-        wframes = None
-        meshes = None
-        if update_coordinates:
-            wframes = self.get_outlines()
-            meshes = self.get_meshes()
-        surf_args = dict.fromkeys(self.permutations)
-        wfrm_args = dict.fromkeys(self.permutations)
-
-        for key, val in sorted(vslices.items()):
-            if update_coordinates:
-                perm = self.permutations[key]
-                surf_args[key] = np.ones_like(meshes[key][perm[0]]) * \
-                    val["loc"]
-                wfrm_args[key] = np.ones_like(wframes[key][perm[0]]) * \
-                    val["loc"]
-                for p in perm:
-                    surf_args[p] = meshes[key][p]
-                    wfrm_args[p] = wframes[key][p]
-
-                self.wireframes[key] = ipv.plot_wireframe(**wfrm_args,
-                                                          color="red")
-                self.wireframes[key].visible = False
-                self.surfaces[key] = ipv.plot_surface(**surf_args)
-                self.members["wireframes"][key] = \
-                    self.wireframes[key]
-                self.members["surfaces"][key] = self.surfaces[key]
-
-            self.surfaces[key].color = self.scalar_map.to_rgba(
-                self.check_transpose(val["slice"]).flatten())
-
-        return
-
-    # Define function to update wireframes
-    def update_slice(self, change):
-        if self.buttons[change["owner"].dim].value is None:
-            self.update_cube(update_coordinates=False)
-        else:
-            # Update only one slice
-            # The dimensions to be sliced have been saved in slider_dims
-            # slice_indices = {"x": 0, "y": 1, "z": 2}
-            dim = change["owner"].dim
-            self.lab[dim].value = self.make_slider_label(
-                self.slider_x[self.name][dim], change["new"])
-
-            # Now move slice
-            ax_dim = self.buttons[dim].value.lower()
-            self.wireframes[ax_dim].visible = True
-            setattr(
-                self.wireframes[ax_dim], ax_dim,
-                getattr(self.wireframes[ax_dim], ax_dim) * 0.0 +
-                self.slider_x[self.name][dim].values[change["new"]])
-
-            self.last_changed_slider_dim = dim
-        return
-
-    # Define function to update surfaces
-    def update_surface(self, event):
-        dim = self.last_changed_slider_dim
-        if dim is not None:
-            # Now move slice
-            index = self.slider[dim].value
-            vslice = self.cube[dim, index]
-            ax_dim = self.buttons[dim].value.lower()
-            self.wireframes[ax_dim].visible = False
-
-            setattr(
-                self.surfaces[ax_dim], ax_dim,
-                getattr(self.surfaces[ax_dim], ax_dim) * 0.0 +
-                self.slider_x[self.name][dim].values[index])
-
-            self.surfaces[self.buttons[dim].value.lower()].color = \
-                self.scalar_map.to_rgba(
-                    self.check_transpose(vslice).flatten())
-        return
-
-    def check_transpose(self, vslice):
-        # Check if dimensions of arrays agree, if not, plot the transpose
-        button_values = [
-            self.buttons[dim].value.lower() for dim in vslice.dims
+        # Find spatial and value limits
+        self.xminmax, self.center_of_mass = self.get_spatial_extents()
+        self.vminmax = [
+            sc.min(self.data_array.data).value,
+            sc.max(self.data_array.data).value
         ]
-        values = vslice.values
-        if ord(button_values[0]) > ord(button_values[1]):
-            values = values.T
-        return values
 
-    def update_showhide(self, owner):
-        owner.value = not owner.value
-        owner.description = "hide" if owner.value else "show"
-        owner.button_style = "success" if owner.value else "danger"
-        ax_dim = self.buttons[owner.dim].value.lower()
-        self.surfaces[ax_dim].visible = owner.value
+        # Create the point cloud with pythreejs
+        self.points_geometry, self.points_material, self.points = \
+            self.create_points_geometry()
+
+        # Create outline around point positions
+        self.outline, self.axticks = self.create_outline()
+
+        # Save the size of the outline box for later
+        self.box_size = np.diff(list(self.xminmax.values()), axis=1).ravel()
+
+        # Define camera: look at the centre of mass of the points
+        camera_lookat = self.center_of_mass
+        camera_pos = np.array(self.center_of_mass) + 1.2 * self.box_size
+        self.camera = p3.PerspectiveCamera(position=list(camera_pos),
+                                           aspect=config.plot.width /
+                                           config.plot.height)
+
+        # Add red/green/blue axes helper
+        self.axes_3d = p3.AxesHelper(10.0 * np.linalg.norm(camera_pos))
+
+        # Create the pythreejs scene
+        self.scene = p3.Scene(children=[
+            self.camera, self.axes_3d, self.points, self.outline, self.axticks
+        ],
+                              background=background)
+
+        # Add camera controller
+        self.controller = p3.OrbitControls(controlling=self.camera,
+                                           target=camera_lookat)
+        self.camera.lookAt(camera_lookat)
+
+        # Render the scene into a widget
+        self.renderer = p3.Renderer(camera=self.camera,
+                                    scene=self.scene,
+                                    controls=[self.controller],
+                                    width=config.plot.width,
+                                    height=config.plot.height)
+
+        # Opacity slider: top value controls opacity if no cut surface is
+        # active. If a cut curface is present, the upper slider is the opacity
+        # of the slice, while the lower slider value is the opacity of the
+        # data not in the cut surface.
+        self.opacity_slider = widgets.FloatRangeSlider(min=0.0,
+                                                       max=1.0,
+                                                       value=[0.1, 1],
+                                                       step=0.01,
+                                                       description="Opacity",
+                                                       continuous_update=True)
+        self.opacity_slider.observe(self.update_opacity, names="value")
+        self.opacity_checkbox = widgets.Checkbox(
+            value=self.opacity_slider.continuous_update,
+            tooltip="Continuous update",
+            indent=False,
+            layout={"width": "initial"})
+        self.opacity_checkbox_link = widgets.jslink(
+            (self.opacity_checkbox, 'value'),
+            (self.opacity_slider, 'continuous_update'))
+
+        # Add buttons to provide a choice of different cut surfaces:
+        # - Cartesian X, Y, Z
+        # - Cylindrical X, Y, Z (cylinder major axis)
+        # - Sperical R
+        # - Value-based iso-surface
+        self.cut_surface_buttons = widgets.ToggleButtons(
+            options=[('X ', 0), ('Y ', 1), ('Z ', 2), ('R ', 6), (' X ', 3),
+                     (' Y ', 4), (' Z ', 5), ('', 7)],
+            value=None,
+            description='Cut surface:',
+            button_style='',
+            tooltips=[
+                'X-plane', 'Y-plane', 'Z-plane', 'Sphere', 'Cylinder-X',
+                'Cylinder-Y', 'Cylinder-Z', 'Value'
+            ],
+            icons=(['cube'] * 3) + ['circle-o'] + (['toggle-on'] * 3) +
+            ['magic'],
+            style={"button_width": "55px"},
+            layout={'width': '350px'})
+        self.cut_surface_buttons.observe(self.update_cut_surface_buttons,
+                                         names="value")
+        # Add a capture for a click event: if the active button is clicked,
+        # this resets the togglebuttons value to None and deletes the cut
+        # surface.
+        self.cut_surface_buttons.on_msg(self.check_if_reset_needed)
+
+        # Add slider to control position of cut surface
+        self.cut_slider = widgets.FloatSlider(min=0,
+                                              max=1,
+                                              disabled=True,
+                                              value=0.5,
+                                              layout={"width": "300px"})
+        self.cut_checkbox = widgets.Checkbox(value=True,
+                                             tooltip="Continuous update",
+                                             indent=False,
+                                             layout={"width": "initial"},
+                                             disabled=True)
+        self.cut_checkbox_link = widgets.jslink(
+            (self.cut_checkbox, 'value'),
+            (self.cut_slider, 'continuous_update'))
+        self.cut_slider.observe(self.update_cut_surface, names="value")
+
+        # Allow to change the thickness of the cut surface
+        self.cut_surface_thickness = widgets.FloatText(
+            value=0.05 * self.box_size.max(),
+            layout={"width": "50px"},
+            disabled=True)
+
+        # Put widgets into boxes
+        self.cut_surface_controls = widgets.HBox([
+            self.cut_surface_buttons, self.cut_slider, self.cut_checkbox,
+            self.cut_surface_thickness
+        ])
+
+        self.box = widgets.VBox([
+            widgets.HBox([self.renderer, self.cbar_image]),
+            widgets.VBox(self.vbox),
+            widgets.HBox([self.opacity_slider, self.opacity_checkbox]),
+            self.cut_surface_controls
+        ])
+
+        # Update list of members to be returned in the SciPlot object
+        self.members.update({
+            "camera": self.camera,
+            "scene": self.scene,
+            "renderer": self.renderer
+        })
+
         return
 
-    def get_outlines(self):
-        outlines = dict()
-        for key, val in self.permutations.items():
-            outlines[key] = dict()
-            outlines[key][val[0]], outlines[key][val[1]] = np.meshgrid(
-                self.xminmax[self.button_axis_to_dim[val[0]]],
-                self.xminmax[self.button_axis_to_dim[val[1]]],
-                indexing="ij")
-        return outlines
+    def create_points_geometry(self):
+        """
+        Make a PointsGeometry using pythreejs
+        """
+        points_geometry = p3.BufferGeometry(
+            attributes={
+                'position': p3.BufferAttribute(array=self.positions),
+                'rgba_color': p3.BufferAttribute(array=self.slice_data(None))
+            })
+        points_material = self.create_points_material()
+        points = p3.Points(geometry=points_geometry, material=points_material)
+        return points_geometry, points_material, points
 
-    def get_meshes(self):
-        meshes = dict()
-        for key, val in self.permutations.items():
-            meshes[key] = dict()
-            meshes[key][val[0]], meshes[key][val[1]] = np.meshgrid(
-                self.slider_x[self.name][self.button_axis_to_dim[
-                    val[0]]].values,
-                self.slider_x[self.name][self.button_axis_to_dim[
-                    val[1]]].values,
-                indexing="ij")
-        return meshes
+    def create_points_material(self):
+        """
+        Define custom raw shader for point cloud to allow to RGBA color format.
+        """
+        return p3.ShaderMaterial(vertexShader='''
+precision highp float;
+attribute vec4 rgba_color;
+varying vec3 mypos;
+varying vec4 vColor;
+varying vec4 projected;
+float xDelta, yDelta, zDelta, delta;
+void main(){
+    vColor = rgba_color;
+    gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.0);
+    // use the delta between the point position and camera position for size
+    xDelta = (position[0]-cameraPosition[0]) * (position[0]-cameraPosition[0]);
+    yDelta = (position[1]-cameraPosition[1]) * (position[1]-cameraPosition[1]);
+    zDelta = (position[2]-cameraPosition[2]) * (position[2]-cameraPosition[2]);
+    delta = pow(xDelta + yDelta + zDelta, 0.5);
+    gl_PointSize = 300.0 / delta;
+}
+''',
+                                 fragmentShader='''
+precision highp float;
+varying vec4 vColor;
+void main() {
+    gl_FragColor = vColor;
+}
+''',
+                                 vertexColors='VertexColors',
+                                 transparent=True,
+                                 depthTest=False,
+                                 alphaTest=0.5)
 
-    def set_axes_range(self):
-        if self.aspect == "equal":
-            max_size = 0.0
-            dx = {"x": 0, "y": 0, "z": 0}
-            for ax in dx.keys():
-                dx[ax] = np.ediff1d(self.xminmax[self.button_axis_to_dim[ax]])
-            max_size = np.amax(list(dx.values()))
-            arrays = dict()
-            for ax, size in dx.items():
-                diff = max_size - size
-                arrays[ax] = [
-                    self.xminmax[self.button_axis_to_dim[ax]][0] - 0.5 * diff,
-                    self.xminmax[self.button_axis_to_dim[ax]][1] + 0.5 * diff
-                ]
+    def get_spatial_extents(self):
+        """
+        Find extents of points in 3D
+        """
+        xminmax = {}
+        for i, x in enumerate('xyz'):
+            xminmax[x] = [
+                np.amin(self.positions[:, i]) - 0.5 * self.pixel_size,
+                np.amax(self.positions[:, i]) + 0.5 * self.pixel_size
+            ]
+        center_of_mass = [
+            0.5 * np.sum(xminmax['x']), 0.5 * np.sum(xminmax['y']),
+            0.5 * np.sum(xminmax['z'])
+        ]
+        return xminmax, center_of_mass
 
-            outl_x, outl_y, outl_z = np.meshgrid(arrays["x"],
-                                                 arrays["y"],
-                                                 arrays["z"],
-                                                 indexing="ij")
-        elif self.aspect == "auto":
-            outl_x, outl_y, outl_z = np.meshgrid(
-                self.xminmax[self.button_axis_to_dim["x"]],
-                self.xminmax[self.button_axis_to_dim["y"]],
-                self.xminmax[self.button_axis_to_dim["z"]],
-                indexing="ij")
+    def create_outline(self):
+        """
+        Make a wireframe cube with tick labels
+        """
+
+        box_geometry = p3.BoxBufferGeometry(
+            self.xminmax['x'][1] - self.xminmax['x'][0],
+            self.xminmax['y'][1] - self.xminmax['y'][0],
+            self.xminmax['z'][1] - self.xminmax['z'][0])
+        edges = p3.EdgesGeometry(box_geometry)
+        outline = p3.LineSegments(
+            geometry=edges,
+            material=p3.LineBasicMaterial(color='#000000'),
+            position=[
+                0.5 * np.sum(self.xminmax['x']),
+                0.5 * np.sum(self.xminmax['y']),
+                0.5 * np.sum(self.xminmax['z'])
+            ])
+
+        axticks = self.generate_axis_ticks()
+
+        return outline, axticks
+
+    def make_axis_tick(self, string, position, color="black", size=1.0):
+        """
+        Make a text-based sprite for axis tick
+        """
+        sm = p3.SpriteMaterial(map=p3.TextTexture(string=string,
+                                                  color=color,
+                                                  size=300,
+                                                  squareTexture=True),
+                               transparent=True)
+        return p3.Sprite(material=sm,
+                         position=position,
+                         scaleToTexture=True,
+                         scale=[size, size, size])
+
+    def generate_axis_ticks(self):
+        """
+        Create ticklabels on outline edges
+        """
+
+        max_extent = np.amax(
+            np.diff(list(self.xminmax.values()), axis=1).ravel())
+        tick_size = 0.1 * max_extent
+        axticks = p3.Group()
+        iden = np.identity(3, dtype=np.float32)
+        ticker = mpl.ticker.MaxNLocator(5)
+        offsets = {
+            'x': [0, self.xminmax['y'][0], self.xminmax['z'][0]],
+            'y': [self.xminmax['x'][0], 0, self.xminmax['z'][0]],
+            'z': [self.xminmax['x'][0], self.xminmax['y'][0], 0]
+        }
+
+        for axis, x in enumerate('xyz'):
+            ticks = ticker.tick_values(self.xminmax[x][0], self.xminmax[x][1])
+            for i in range(2, len(ticks)):
+                tick_pos = iden[axis] * ticks[i] + offsets[x]
+                axticks.add(
+                    self.make_axis_tick(string=value_to_string(ticks[i],
+                                                               precision=1),
+                                        position=tick_pos.tolist(),
+                                        size=tick_size))
+        # Add bottom corner
+        tick_pos = iden[0] * ticks[1] + offsets['x']
+        axticks.add(
+            self.make_axis_tick(string=value_to_string(ticks[1], precision=1),
+                                position=tick_pos.tolist(),
+                                size=tick_size))
+
+        return axticks
+
+    def create_colorbar(self):
+        """
+        Make image from matplotlib colorbar.
+        """
+        height_inches = config.plot.height / config.plot.dpi
+        fig = mpl.figure.Figure(figsize=(height_inches * 0.2, height_inches),
+                                dpi=config.plot.dpi)
+        canvas = backend_agg.FigureCanvasAgg(fig)
+        ax = fig.add_axes([0.05, 0.02, 0.25, 0.96])
+        cb1 = mpl.colorbar.ColorbarBase(
+            ax,
+            cmap=cm.get_cmap(self.params["values"][self.name]["cmap"]),
+            norm=self.params["values"][self.name]["norm"])
+        cb1.set_label(name_with_unit(var=self.data_array, name=""))
+        canvas.draw()
+        image = np.frombuffer(canvas.tostring_rgb(), dtype='uint8')
+        shp = list(fig.canvas.get_width_height())[::-1] + [3]
+        self.cbar_image.value = pil.Image.fromarray(
+            image.reshape(shp))._repr_png_()
+
+    def update_opacity(self, change):
+        """
+        Update opacity of all points when opacity slider is changed.
+        Take cut surface into account if present.
+        """
+        if self.cut_surface_buttons.value is None:
+            arr = self.points_geometry.attributes["rgba_color"].array
+            arr[:, 3] = change["new"][1]
+            self.points_geometry.attributes["rgba_color"].array = arr
         else:
-            raise RuntimeError("Unknown aspect ratio: {}".format(self.aspect))
+            self.update_cut_surface({"new": self.cut_slider.value})
 
-        self.fig.xlim = list(outl_x.flatten()[[0, -1]])
-        self.fig.ylim = list(outl_y.flatten()[[0, -1]])
-        self.fig.zlim = list(outl_z.flatten()[[0, -1]])
+    def check_if_reset_needed(self, owner, content, buffers):
+        if owner.value == self.current_cut_surface_value:
+            self.cut_surface_buttons.value = None
+        self.current_cut_surface_value = owner.value
+
+    def update_cut_surface_buttons(self, change):
+        if change["new"] is None:
+            self.cut_slider.disabled = True
+            self.cut_checkbox.disabled = True
+            self.cut_surface_thickness.disabled = True
+            self.update_opacity({"new": self.opacity_slider.value})
+        else:
+            if change["old"] is None:
+                self.cut_slider.disabled = False
+                self.cut_checkbox.disabled = False
+                self.cut_surface_thickness.disabled = False
+            self.update_cut_slider_bounds()
+
+    def update_cut_slider_bounds(self):
+        # Cartesian X, Y, Z
+        if self.cut_surface_buttons.value < 3:
+            minmax = self.xminmax["xyz" [self.cut_surface_buttons.value]]
+            if minmax[0] < self.cut_slider.max:
+                self.cut_slider.min = minmax[0]
+                self.cut_slider.max = minmax[1]
+            else:
+                self.cut_slider.max = minmax[1]
+                self.cut_slider.min = minmax[0]
+            self.cut_slider.value = 0.5 * (minmax[0] + minmax[1])
+        # Cylindrical X, Y, Z
+        elif self.cut_surface_buttons.value < 6:
+            j = self.cut_surface_buttons.value - 3
+            remaining_axes = self.permutations["xyz" [j]]
+            self.remaining_inds = [(j + 1) % 3, (j + 2) % 3]
+            rmax = np.abs([
+                self.xminmax[remaining_axes[0]][0],
+                self.xminmax[remaining_axes[1]][0],
+                self.xminmax[remaining_axes[0]][1],
+                self.xminmax[remaining_axes[1]][1]
+            ]).max()
+            self.cut_slider.min = 0
+            self.cut_slider.max = rmax * np.sqrt(2.0)
+            self.cut_slider.value = 0.5 * self.cut_slider.max
+        # Spherical
+        elif self.cut_surface_buttons.value == 6:
+            rmax = np.abs(list(self.xminmax.values())).max()
+            self.cut_slider.min = 0
+            self.cut_slider.max = rmax * np.sqrt(3.0)
+            self.cut_slider.value = 0.5 * self.cut_slider.max
+        # Value iso-surface
+        elif self.cut_surface_buttons.value == 7:
+            self.cut_slider.min = self.vminmax[0]
+            self.cut_slider.max = self.vminmax[1]
+            self.cut_slider.value = 0.5 * (self.vminmax[0] + self.vminmax[1])
+
+        # Remember to reset surface thickness when changing between spatial
+        # slicing and value slicing
+        if self.cut_surface_buttons.value < 7:
+            self.cut_surface_thickness.value = 0.05 * self.box_size.max()
+        else:
+            self.cut_surface_thickness.value = 0.05 * (self.vminmax[1] -
+                                                       self.vminmax[0])
+
+        # Update slider step to avoid too fine granularity which slows down
+        # interaction.
+        self.cut_slider.step = (self.cut_slider.max -
+                                self.cut_slider.min) / self.cut_slider_steps
+
+    def update_cut_surface(self, change):
+        newc = None
+        # Cartesian X, Y, Z
+        if self.cut_surface_buttons.value < 3:
+            newc = np.where(
+                np.abs(self.positions[:, self.cut_surface_buttons.value] -
+                       change["new"]) < self.cut_surface_thickness.value,
+                self.opacity_slider.upper, self.opacity_slider.lower)
+        # Cylindrical X, Y, Z
+        elif self.cut_surface_buttons.value < 6:
+            newc = np.where(
+                np.abs(
+                    np.sqrt(self.positions[:, self.remaining_inds[0]] *
+                            self.positions[:, self.remaining_inds[0]] +
+                            self.positions[:, self.remaining_inds[1]] *
+                            self.positions[:, self.remaining_inds[1]]) -
+                    change["new"]) < self.cut_surface_thickness.value,
+                self.opacity_slider.upper, self.opacity_slider.lower)
+        # Spherical
+        elif self.cut_surface_buttons.value == 6:
+            newc = np.where(
+                np.abs(
+                    np.sqrt(self.positions[:, 0] * self.positions[:, 0] +
+                            self.positions[:, 1] * self.positions[:, 1] +
+                            self.positions[:, 2] * self.positions[:, 2]) -
+                    change["new"]) < self.cut_surface_thickness.value,
+                self.opacity_slider.upper, self.opacity_slider.lower)
+        # Value iso-surface
+        elif self.cut_surface_buttons.value == 7:
+            newc = np.where(
+                np.abs(self.vslice - change["new"]) <
+                self.cut_surface_thickness.value, self.opacity_slider.upper,
+                self.opacity_slider.lower)
+
+        # Unfortunately, one cannot edit the value of the geometry array
+        # in-place, as this does not trigger an update on the threejs side.
+        # We have to update the entire array.
+        c3 = self.points_geometry.attributes["rgba_color"].array
+        c3[:, 3] = newc
+        self.points_geometry.attributes["rgba_color"].array = c3
+
+    def slice_data(self, change):
+        """
+        Slice the extra dimensions down and update the slice values
+        """
+        self.vslice = self.data_array
+        if self.params["masks"][self.name]["show"]:
+            mslice = self.masks
+        # Slice along dimensions with active sliders
+        for dim, val in self.slider.items():
+            if not val.disabled:
+                self.lab[dim].value = self.make_slider_label(
+                    self.slider_x[self.name][dim], val.value)
+                self.vslice = self.vslice[val.dim, val.value]
+                # At this point, after masks were combined, all their
+                # dimensions should be contained in the data_array.dims.
+                if self.params["masks"][
+                        self.name]["show"] and dim in mslice.dims:
+                    mslice = mslice[val.dim, val.value]
+
+        if self.params["masks"][self.name]["show"]:
+            # Use automatic broadcasting in Scipp variables
+            msk = sc.Variable(dims=self.vslice.dims,
+                              values=np.ones(self.vslice.shape,
+                                             dtype=np.int32))
+            msk *= sc.Variable(dims=mslice.dims,
+                               values=mslice.values.astype(np.int32))
+            msk = msk.values
+
+        self.vslice = self.vslice.values.flatten()
+        colors = self.scalar_map.to_rgba(self.vslice).astype(np.float32)
+
+        if self.params["masks"][self.name]["show"]:
+            masks_inds = np.where(msk.flatten())
+            masks_colors = self.masks_scalar_map.to_rgba(
+                self.vslice[masks_inds]).astype(np.float32)
+            colors[masks_inds] = masks_colors
+
+        return colors
+
+    def update_slice(self, change):
+        """
+        Update colors of points.
+        """
+        new_colors = self.slice_data(change)
+        new_colors[:, 3] = self.points_geometry.attributes[
+            "rgba_color"].array[:, 3]
+        self.points_geometry.attributes["rgba_color"].array = new_colors
+        return
+
+    def toggle_masks(self, change):
+        """
+        Show/hide masks
+        """
+        self.params["masks"][self.name]["show"] = change["new"]
+        change["owner"].description = "Hide masks" if change["new"] else \
+            "Show masks"
+        self.update_slice(None)
         return
